+++ src/main/java/com/part4/team09/otboo/module/domain/feed/exception/FeedErrorCode.java
@@ -5,7 +5,8 @@ import org.springframework.http.HttpStatus;
 
 public enum FeedErrorCode implements ErrorCode {
 
-  FEED_NOT_FOUND(HttpStatus.NOT_FOUND, "피드를 찾을 수 없습니다.");
+  FEED_NOT_FOUND(HttpStatus.NOT_FOUND, "피드를 찾을 수 없습니다."),
+  LIKE_NOT_FOUND(HttpStatus.NOT_FOUND, "좋아요를 찾을 수 없습니다.");
 
   private HttpStatus httpStatus;
   private String message;
