+++ src/main/java/com/part4/team09/otboo/module/domain/feed/service/LikeService.java
@@ -1,15 +1,14 @@
 package com.part4.team09.otboo.module.domain.feed.service;
 
 import com.part4.team09.otboo.module.domain.feed.dto.FeedDto;
-import com.part4.team09.otboo.module.domain.feed.entity.Feed;
 import com.part4.team09.otboo.module.domain.feed.entity.Like;
-import com.part4.team09.otboo.module.domain.feed.exception.FeedNotFoundException;
+import com.part4.team09.otboo.module.domain.feed.exception.feed.FeedNotFoundException;
+import com.part4.team09.otboo.module.domain.feed.exception.like.LikeNotFoundException;
 import com.part4.team09.otboo.module.domain.feed.mapper.FeedDtoAssembler;
 import com.part4.team09.otboo.module.domain.feed.repository.FeedRepository;
 import com.part4.team09.otboo.module.domain.feed.repository.LikeRepository;
 import com.part4.team09.otboo.module.domain.user.exception.UserNotFoundException;
 import com.part4.team09.otboo.module.domain.user.repository.UserRepository;
-import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -46,10 +45,9 @@ public class LikeService {
     likeRepository.deleteById(like.getId());
   }
 
-  // TODO: 커스텀 예외 만들기
   private Like getLikeOrThrow(UUID userId, UUID feedId) {
     return likeRepository.findByUserIdAndFeedId(userId, feedId)
-        .orElseThrow(() -> FeedNotFoundException.withId(feedId));
+        .orElseThrow(() -> LikeNotFoundException.withId(userId, feedId));
   }
 
   public boolean isLikedByMe(UUID userId, UUID feedId) {
